Moved menu into pages
